chore: Renames token table stylesheet
